+++ CHANGES.rst
@@ -1,6 +1,9 @@
 v34.4.0
 -------
 
+* #995: In MSVC support, add support for Microsoft
+  Build Tools 2017.
+
 * #999 via #1007: Extend support for declarative package
   config in a setup.cfg file to include the options
   ``python_requires`` and ``py_modules``.